FP8 Contribution
I developed the feature where user can select a particular dance among the list of dances in a category or region that the dance belongs to and be navigated to the page to view the dance. The dance page displays the dance to be able to be read, deleted (completed), and updated (in-progress), when the user clicks on the dance.